+++ perl/FSM/Support/ISFPublicInterfaceContract.pm
@@ -443,6 +443,7 @@ sub build_isf_public_interface_contract {
             't/1274-isf-enum-member-rule-guard-values.t',
             't/1275-isf-enum-member-condition-values.t',
             't/1276-isf-enum-member-activation-aggregate-params.t',
+            't/1277-isf-enum-member-actor-aggregate-params.t',
         ],
         guidance => [
             'Treat this as the first bounded public ISF downstream-consumer contract, advertised through embedding.isf_public_interface.',